Introduce "Object Names"
Adds a new attribute that can optionally be used to tell UIOMatic which Property is meant to be the Name of your object

When specified, the object name will be editable from the header of the screen, just like when editing Umbraco content nodes.  Also, specifying the Name field will cause it to be the linkable column in the list view.